back/app/requirements.txt
flask
pillow
ultralytics
flask-cors
fastapi
python-multipart
uvicorn